requirements.txt
dotenv
openai
pydantic
matplotlib==3.9.0
google-adk
PyQt6
